Filter and search seperated with the logic
