--- components/NavigationBar/Layout.js
@@ -1,9 +1,0 @@
-import MainHeader from "./MainHeader";
-
-export default function Layout({children}){
-    <>
-        <Header></Header>
-        <main>{children}</main>
-    </>
-
-}
